Fixes and refactor build functions into build.ts
