+++ templates/partials/nav.html
@@ -96,6 +96,10 @@
                     <a href="{% url 'map' %}">Map</a>
                 </li>
 
+                <li class="p-4 pb-5 {% active_link 'text-purple' 'remix' %}">
+                    <a href="{% url 'remix' %}">Remix</a>
+                </li>
+
                 {% if user.is_authenticated %}
                     <li class="p-4 pb-5 {% active_link 'text-purple' 'account_all_chats' 'user_chat' %}">
                         <a href="{% url 'account_all_chats' %}">Messages</a>
