+++ templates/partials/overlay-river-closable.html
@@ -2,7 +2,7 @@
 
 {% block overlay_header %}
     <div class="flex justify-between">
-        <h2 class="text-large text-center">
+        <h2 class="text-large text-center flex-grow">
             {% block overlay_title %}{% endblock %}
         </h2>
         <button class="cursor-pointer inline-block" @click="overlayClose()">
