+++ templates/partials/vertical-spacer.html
@@ -1,1 +1,1 @@
-<div class="h-{{ space }}"></div>
+<div class="h-{{ space }} shrink-0"></div>
