--- templates/ts/directives/remix/models/.gitignore
@@ -1,1 +1,0 @@
-*.glb
